Allow building generated code with OpenMP and MSVC

+++ src/include/souffle/datastructure/LambdaBTree.h
@@ -314,7 +314,7 @@ public:
                 // split this node
                 auto old_root = this->root;
                 idx -= cur->rebalance_or_split(
-                        const_cast<typename parenttype::node**>(&this->root), this->root_lock, idx, parents);
+                        const_cast<typename parenttype::node**>(&this->root), this->root_lock, static_cast<int>(idx), parents);
 
                 // release parent lock
                 for (auto it = parents.rbegin(); it != parents.rend(); ++it) {
